Another big commit: * KGet ported to qt4. Here there are still some issues that need to be resolved. For example I had to disable the bittorrent plugin becouse currently I don't know hot to make it compile again. * Started implementing a new sidebar in qt3. Then I switched to qt4 and realized I had to rewrite it, since I made use of the QListBox class :-( So I did, this time using the QVBoxLayout class. More to come..
svn path=/branches/work/make_kget_cool/kget/; revision=442517

+++ core/transfer.cpp
@@ -12,7 +12,9 @@
 #include <klocale.h>
 #include <kiconloader.h>
 
-#include <qdom.h>
+#include <QDomElement>
+//Added by qt3to4:
+#include <QPixmap>
 
 #include "core/transfer.h"
 #include "core/transferhandler.h"
@@ -73,8 +75,8 @@ void Transfer::save(QDomElement e)
 {
     e.setAttribute("Source", m_source.url());
     e.setAttribute("Dest", m_dest.url());
-    e.setAttribute("TotalSize", m_totalSize);
-    e.setAttribute("ProcessedSize", m_processedSize);
+    e.setAttribute("TotalSize", (qulonglong) m_totalSize);
+    e.setAttribute("ProcessedSize", (qulonglong) m_processedSize);
 }
 
 void Transfer::load(QDomElement e)
